+++ ios/ReactNativeFlurry/ReactNativeFlurry.m
@@ -57,7 +57,7 @@
 #endif
 
 static NSString * const originName = @"react-native-flurry-sdk";
-static NSString * const originVersion = @"5.0.0";
+static NSString * const originVersion = @"5.1.0";
 
 @interface ReactNativeFlurry ()<RNFlurryEventDispatcherDelegate>